Build NLP variable bounds as float arrays
get_variable_bounds() did `np.array(var_lower)`, which infers an integer
dtype when every bound is integer-valued -- e.g. a variable with
nonneg=True contributes a lower bound of 0. The KNITRO interface then
does `lb[lb == -np.inf] = -knitro.KN_INFINITY`; KN_INFINITY (~1e20)
does not fit in int64, raising `OverflowError: Python int too large to
convert to C long`.

This surfaced once the over-broad UNO skips were scoped down, exposing
test_geo_mean[KNITRO] / test_geo_mean2[KNITRO]. Bounds are real-valued,
so construct lb/ub with dtype=float (constraint bounds cl/cu already are).

+++ cvxpy/reductions/solvers/nlp_solvers/nlp_solver.py
@@ -136,8 +136,11 @@ class Bounds:
             ub_flat = np.asarray(ub).flatten(order='F')
             var_lower.extend(lb_flat)
             var_upper.extend(ub_flat)
-        self.lb = np.array(var_lower)
-        self.ub = np.array(var_upper)
+        # Force float dtype: all-integer bounds (e.g. nonneg=True gives lb 0)
+        # would otherwise infer an integer array, and replacing entries with a
+        # solver's infinity sentinel (~1e20) overflows int64.
+        self.lb = np.array(var_lower, dtype=float)
+        self.ub = np.array(var_upper, dtype=float)
 
     def construct_initial_point(self) -> None:
         """ Loop through all variables and collect the intial point."""
